portal developed; small fix css vars
