+++ src/app/styles/index.scss
@@ -3,11 +3,15 @@
 @import 'themes/dark';
 @import 'themes/light';
 
-.app {
+body {
   font: var(--font-m);
+  color: var(--primary-color);
+}
+
+.app {
+  min-height: 100vh;
   background: var(--bg-color);
   color: var(--primary-color);
-  min-height: 100vh;
 }
 
 .page-wrapper {
