+++ src/app/styles/themes/dark.scss
@@ -1,8 +1,8 @@
 .app.dark {
   --bg-color: #0e0e75;
   --inverted-bg-color: #d2d2d8;
-  --primary-color: rgb(8 209 8);
-  --secondary-color: rgb(6 114 6);
-  --inverted-primary-color: rgb(2 29 138);
-  --inverted-secondary-color: gb(8, 79, 231);
+  --primary-color: #08d108;
+  --secondary-color: #067206;
+  --inverted-primary-color: #021d8a;
+  --inverted-secondary-color: #084fe7;
 }
