+++ src/app/styles/themes/light.scss
@@ -1,8 +1,8 @@
 .app.light {
   --bg-color: #d2d2d8;
   --inverted-bg-color: #0e0e75;
-  --primary-color: rgb(2 29 138);
-  --secondary-color: rgb(8 79 231);
-  --inverted-primary-color: rgb(8 209 8);
-  --inverted-secondary-color: rgb(6 114 6);
+  --primary-color: #021d8a;
+  --secondary-color: #084fe7;
+  --inverted-primary-color: #08d108;
+  --inverted-secondary-color: #067206;
 }
