+++ src/index.tsx
@@ -2,6 +2,7 @@ import { render } from 'react-dom'
 import App from './app/App'
 import { BrowserRouter } from 'react-router-dom'
 import { ThemeProvider } from 'app/providers/ThemeProvider'
+import 'app/styles/index.scss'
 
 import './shared/config/i18n/i18n'
 import { ErrorBoundary } from 'app/providers/ErrorBoundary'
