+++ src/shared/ui/Modal/Modal.module.scss
@@ -7,6 +7,18 @@
   z-index: var(--modal-z-index);
   opacity: 0;
   pointer-events: none;
+  visibility: hidden;
+  transition: all 0.5s ease;
+}
+
+.modal.dark {
+  --inverted-primary-color: #021d8a;
+  --inverted-bg-color: #d2d2d8;
+}
+
+.modal.light {
+  --inverted-primary-color: #08d108;
+  --inverted-bg-color: #0e0e75;
 }
 
 .mask {
@@ -24,14 +36,13 @@
   background: var(--inverted-bg-color);
   color: var(--inverted-primary-color);
   max-width: 50%;
-  transition: 0.4s transform;
-  transform: scale(0.5);
 }
 
 .opened {
   pointer-events: auto;
   opacity: 1;
   z-index: var(--modal-z-index);
+  visibility: visible;
 
   .content {
     transform: scale(1);
