+++ src/shared/ui/Modal/Modal.tsx
@@ -1,6 +1,8 @@
 import { ReactNode, useCallback, useEffect, useRef, useState } from 'react'
 import cls from './Modal.module.scss'
 import { classNames } from 'shared/lib/classNames/classNames'
+import { Portal } from '../Portal/Portal'
+import { useTheme } from '../../../app/providers/ThemeProvider'
 
 interface ModalProps {
   children?: ReactNode
@@ -19,18 +21,16 @@ export const Modal = ({
 }: ModalProps): JSX.Element => {
   const [isClosing, setIsClosing] = useState(false)
   const timerRef = useRef<ReturnType<typeof setTimeout>>()
+  const { theme } = useTheme()
   const mods: Record<string, boolean> = {
     [cls.opened]: isOpen,
     [cls.isClosing]: isClosing,
+    [cls[theme]]: true,
   }
 
   const closeHandler = useCallback(() => {
     if (onClose) {
-      setIsClosing(true)
-      timerRef.current = setTimeout(() => {
-        onClose()
-        setIsClosing(false)
-      }, ANIMATION_DELAY)
+      onClose()
     }
   }, [onClose])
 
@@ -58,12 +58,14 @@ export const Modal = ({
   }, [isOpen, onEscKeyDown])
 
   return (
-    <div className={classNames(cls.modal, mods, [className])}>
-      <div className={cls.mask} onClick={closeHandler}>
-        <div className={cls.content} onClick={onContentClick}>
-          {children}
+    <Portal>
+      <div className={classNames(cls.modal, mods, [className])}>
+        <div className={cls.mask} onClick={closeHandler}>
+          <div className={cls.content} onClick={onContentClick}>
+            {children}
+          </div>
         </div>
       </div>
-    </div>
+    </Portal>
   )
 }
