+++ src/shared/ui/Portal/Portal.tsx
@@ -1,0 +1,11 @@
+import { ReactNode } from 'react'
+import { createPortal } from 'react-dom'
+
+interface PortalProps {
+  children: ReactNode
+  element?: Element
+}
+
+export const Portal = ({ children, element = document.body }: PortalProps) => {
+  return createPortal(children, element)
+}
